src/requirements.txt
numpy==1.22.3
pillow==9.1.0
matplotlib==3.5.1